Cleaner traits for fly::Json
