+++ fly/traits/traits.h
@@ -1,16 +1,6 @@
 #pragma once
 
-#include <array>
-#include <deque>
-#include <forward_list>
-#include <list>
-#include <map>
-#include <set>
 #include <type_traits>
-#include <unordered_map>
-#include <unordered_set>
-#include <utility>
-#include <vector>
 
 /**
  * Define SFINAE tests for whether a function is declared for a type.
@@ -120,210 +110,5 @@ using enable_if_none =
 template <typename... Conditions>
 using enable_if_not_all =
     std::enable_if_t<std::negation_v<std::conjunction<Conditions...>>, bool>;
-
-/**
- * Tests for whether a given type is a signed integer.
- */
-struct if_signed_integer
-{
-    struct __
-    {
-        template <typename T>
-        inline static constexpr bool is_signed_integer = std::bool_constant<
-            std::is_integral_v<std::decay_t<T>> &&
-            std::is_signed_v<std::decay_t<T>> &&
-            !std::is_same_v<bool, std::decay_t<T>>>::value;
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_signed_integer<T>, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_signed_integer<T>, bool>;
-};
-
-/**
- * Tests for whether a given type is an unsigned integer.
- */
-struct if_unsigned_integer
-{
-    struct __
-    {
-        template <typename T>
-        inline static constexpr bool is_unsigned_integer = std::bool_constant<
-            std::is_integral_v<std::decay_t<T>> &&
-            std::is_unsigned_v<std::decay_t<T>> &&
-            !std::is_same_v<bool, std::decay_t<T>>>::value;
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_unsigned_integer<T>, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_unsigned_integer<T>, bool>;
-};
-
-/**
- * Tests for whether a given type is a float.
- */
-struct if_floating_point
-{
-    struct __
-    {
-        template <typename T>
-        inline static constexpr bool is_floating_point = std::bool_constant<
-            std::is_floating_point_v<std::decay_t<T>>>::value;
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_floating_point<T>, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_floating_point<T>, bool>;
-};
-
-/**
- * Tests for whether a given type is numeric.
- */
-struct if_numeric
-{
-    struct __
-    {
-        template <typename T>
-        inline static constexpr bool is_numeric = std::bool_constant<
-            std::is_arithmetic_v<std::decay_t<T>> &&
-            !std::is_same_v<bool, std::decay_t<T>>>::value;
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_numeric<T>, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_numeric<T>, bool>;
-};
-
-/**
- * Tests for whether a given type is a boolean.
- */
-struct if_boolean
-{
-    struct __
-    {
-        template <typename T>
-        inline static constexpr bool is_boolean =
-            std::bool_constant<std::is_same_v<bool, std::decay_t<T>>>::value;
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_boolean<T>, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_boolean<T>, bool>;
-};
-
-/**
- * Tests for whether a given type is map-like (i.e. holds key-value pairs).
- */
-struct if_map
-{
-    struct __
-    {
-        template <typename>
-        struct is_map : std::false_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_map<std::map<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_map<std::multimap<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_map<std::unordered_map<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_map<std::unordered_multimap<Args...>> : std::true_type
-        {
-        };
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_map<T>::value, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_map<T>::value, bool>;
-};
-
-/**
- * Tests for whether a given type is array-like (i.e. holds value sequences).
- */
-struct if_array
-{
-    struct __
-    {
-        template <typename>
-        struct is_array : std::false_type
-        {
-        };
-
-        template <typename T, std::size_t N>
-        struct is_array<std::array<T, N>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::deque<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::forward_list<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::list<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::multiset<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::set<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::unordered_multiset<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::unordered_set<Args...>> : std::true_type
-        {
-        };
-
-        template <typename... Args>
-        struct is_array<std::vector<Args...>> : std::true_type
-        {
-        };
-    };
-
-    template <typename T>
-    using enabled = std::enable_if_t<__::is_array<T>::value, bool>;
-
-    template <typename T>
-    using disabled = std::enable_if_t<!__::is_array<T>::value, bool>;
-};
 
 } // namespace fly
